tools/complexity_daemon/requirements.txt
watchdog>=4.0.0
lizard>=1.17.0
gitpython>=3.1.0
typer>=0.12.0
rich>=13.0.0
tomli>=2.0.0
toml>=0.10.2
